requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
whitenoise==6.2.0
